rename scribe.hub -> hub.herald
